delete movies ok

+++ src/main/java/com/Alkemy/Challenge/service/imp/CharacterServiceImpl.java
@@ -2,23 +2,29 @@ package com.Alkemy.Challenge.service.imp;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import org.springframework.stereotype.Service;
 
 import com.Alkemy.Challenge.dto.CharacterBasicDTO;
 import com.Alkemy.Challenge.dto.CharacterDTO;
+import com.Alkemy.Challenge.dto.CharacterFiltersDTO;
 import com.Alkemy.Challenge.entity.CharacterEntity;
+import com.Alkemy.Challenge.exception.ParamNotFound;
 import com.Alkemy.Challenge.mapper.CharacterMapper;
 import com.Alkemy.Challenge.repository.CharacterRepository;
+import com.Alkemy.Challenge.repository.specifications.CharacterSpecification;
 import com.Alkemy.Challenge.service.CharacterService;
 
 import lombok.AllArgsConstructor;
 
 @Service
 @AllArgsConstructor
-public class CharacterServiceImp implements CharacterService{
+public class CharacterServiceImpl implements CharacterService{
 	
 	private final CharacterRepository characterRepository;
+	
+	private final CharacterSpecification characterSpecification;
 
 	private final CharacterMapper characterMapper;
 	
@@ -30,6 +36,14 @@ public class CharacterServiceImp implements CharacterService{
 		
 		
 		return dtoBasicList;
+	}
+	
+	@Override
+	public List<CharacterDTO> getByFilters(String name, String age, Set<Long> movies) {
+		CharacterFiltersDTO filtersDTO = new CharacterFiltersDTO(name, age, movies);
+		List<CharacterEntity> entities = characterRepository.findAll(characterSpecification.getByFilters(filtersDTO));
+		List<CharacterDTO> dtos = characterMapper.characterEntitySet2DTOList(entities, true); 
+		return dtos;
 	}
 
 	@Override
@@ -43,18 +57,26 @@ public class CharacterServiceImp implements CharacterService{
 
 	@Override
 	public void delete(Long id) {
-
+		if(characterRepository.findById(id).isEmpty()){
+			throw new ParamNotFound("Id no encontrado");
+		}
 		characterRepository.deleteById(id);
+		
+		
 	}
 
 	@Override
 	public CharacterDTO update(CharacterDTO dto) {
 		
 		Optional<CharacterEntity> character2Update = characterRepository.findById(dto.getId());
+		if(character2Update.isEmpty()) {
+			throw new ParamNotFound("Id no encontrado");			
+		}
 		characterMapper.characterEntityRefreshValues(character2Update.get(), dto);
 		CharacterEntity characterSaved = characterRepository.save(character2Update.get());
 		
 		return characterMapper.characterEntity2DTO(characterSaved, true);
 	}
+
 
 }
